requirements.txt
httpx
pyperclip
pynput
python-dotenv
